Phase 6E-B19 corrective #2: unsigned shift proof + 6 dedicated tests
func_8005E968 arithmetic:
- PROVED: (packed >> 1) & 0x007F7F7F is identical to MIPS sra+and
  for all 2^32 inputs — the AND mask clears top 9 bits, making
  signedness irrelevant. Replaced (int32_t)cast with portable
  unsigned expression.

New tests (258 total, was 252):
- test_5E968_contract: 0x80808080 → 0x00404040
- test_5E968_boundary_values: 12 corner cases against MIPS model
- test_5F844_a0_zero: 0x395D, 0x84, 0xA4
- test_5F844_a0_nonzero: 0x3A1C, 0xCC, 0xA4
- test_5E588_direct_boot_state: pointer propagation,
  flag bytes, halfwords, gp-relative stores, leaf results,
  DRAWENV/DISPENV spot checks — D_800A21F4/21F0 proven
  overwritten by DRAWENV_C (retail behavior)
- test_5E588_ramreset_rerun: PE_RamReset proof

Gates:
- 258/258 normal + ASan+UBSan
- 3× deterministic: 258/258 identical
- RNG/LZCR/callback/dispatcher oracles: all PASS
- Docker: EXACT SHA-1 MATCH
- git diff --check: clean
- func_8005E588 produces zero bootstrap stubs (tested directly)
- Strict frontier: func_80062568 (next dispatcher callee)

+++ pc_port/game/boot/func_8005E968_port.c
@@ -31,5 +31,5 @@
 void func_8005E968(uint32_t packed)
 {
     PE_StoreU32(GA_COLOR_ORIG, packed);
-    PE_StoreU32(GA_COLOR_HALF, ((int32_t)packed >> 1) & 0x007F7F7Fu);
+    PE_StoreU32(GA_COLOR_HALF, (packed >> 1) & 0x007F7F7Fu);
 }
